1.修复在播放语音时进入其他界面可能出现异常的Bug 2.修复在聊天界面群聊时可能抛出的异常 3.修复在录音时获取音量时可能出现的异常 4.第一次登录后，进入MeInfoActivity中可以刷新头像
